backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.27.0
beautifulsoup4==4.12.3
feedparser==6.0.11
# playwright is optional (heavy). Only enable if you implement real marketplace scraping.
# playwright==1.47.0
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.2
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
